# swagger_documentation.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Drogaleste API",
    "description": "Documentação da API da Drogaleste"
  },
  "host": "apidrogaleste.ddns.net:7150",
  "basePath": "api/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "JWT": {
      "description": "JWT token",
      "type": "Bearer",
      "in": "body",
      "name": "Authorization"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/oauth/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "clientId": {
                  "example": "any"
                },
                "clientSecret": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/drogaleste/product/stock/refresh/{cod}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "cod",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/drogaleste/product/stock": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "produto",
            "in": "header",
            "type": "string"
          },
          {
            "name": "loja",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/drogaleste/product/charge": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "pagenumber",
            "in": "header",
            "type": "string"
          },
          {
            "name": "pagerows",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/drogaleste/product/refreshregistration": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "loja",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/drogaleste/product/{cod}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "cod",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "loja",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/drogaleste/store/{cod}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "cod",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  }
}